ai_model/requirements.txt
numpy
scipy
scikit-learn
matplotlib
pandas
fastapi
uvicorn
gunicorn==19.9.0
uvloop
httptools
tensorflow==2.9.1
load_dotenv